Move GLTransform matrix of all entities in the PhysicalWorldLocation component since the matrix is a representation of a location.
