+++ include/systems/OpenGLSystem.h
@@ -20,7 +20,7 @@
 #include "resources/GLTexture.h"
 #include "components/GLScreenQuad.h"
 #include "Sigma.h"
-#include "components/ControllableMove.h"
+#include "components/PhysicalWorldLocation.h"
 
 struct IGLView;
